LPS-77580 Move UADRegistry to web module
