+++ modules/apps/foundation/user-associated-data/user-associated-data-web/build.gradle
@@ -1,4 +1,7 @@
 dependencies {
+	compileOnly group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
+
+	provided group: "com.liferay", name: "com.liferay.osgi.service.tracker.collections", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.petra.lang", version: "1.0.0"
 	provided group: "com.liferay", name: "com.liferay.petra.string", version: "1.0.0"
 	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
