+++ modules/apps/foundation/user-associated-data/user-associated-data-web/src/main/java/com/liferay/user/associated/data/web/internal/registry/UADRegistry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.user.associated.data.internal.registry;
+package com.liferay.user.associated.data.web.internal.registry;
 
 import com.liferay.osgi.service.tracker.collections.map.ServiceReferenceMapper;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
@@ -23,7 +23,6 @@ import com.liferay.user.associated.data.anonymizer.UADEntityAnonymizer;
 import com.liferay.user.associated.data.display.UADEntityDisplay;
 import com.liferay.user.associated.data.entity.UADEntity;
 import com.liferay.user.associated.data.exporter.UADEntityExporter;
-import com.liferay.user.associated.data.registry.UADRegistry;
 import com.liferay.user.associated.data.util.UADEntitySetComposite;
 import com.liferay.user.associated.data.util.UADEntityTypeComposite;
 
@@ -43,7 +42,7 @@ import org.osgi.service.component.annotations.Component;
 * @author William Newbury
 */
 @Component(immediate = true, service = UADRegistry.class)
-public class UADRegistryImpl implements UADRegistry {
+public class UADRegistry {
 
 	public UADEntityAggregator getUADEntityAggregator(String key) {
 		return _uadEntityAggregatorTrackerMap.getService(key);
@@ -77,17 +76,14 @@ public class UADRegistryImpl implements UADRegistry {
 		return _uadEntityAnonymizerTrackerMap.values();
 	}
 
-	@Override
 	public UADEntityDisplay getUADEntityDisplay(String key) {
 		return _uadEntityDisplayTrackerMap.getService(key);
 	}
 
-	@Override
 	public UADEntityDisplay getUADEntityDisplay(UADEntity uadEntity) {
 		return getUADEntityDisplay(uadEntity.getUADRegistryKey());
 	}
 
-	@Override
 	public Set<String> getUADEntityDisplayKeySet() {
 		return _uadEntityDisplayTrackerMap.keySet();
 	}
@@ -108,7 +104,6 @@ public class UADRegistryImpl implements UADRegistry {
 		return _uadEntityExporterTrackerMap.values();
 	}
 
-	@Override
 	public List<UADEntitySetComposite> getUADEntitySetComposites(long userId) {
 		Map<String, List<UADEntityTypeComposite>> uadEntityTypeCompositesMap =
 			new HashMap<>();
@@ -151,7 +146,6 @@ public class UADRegistryImpl implements UADRegistry {
 		return uadEntitySetComposites;
 	}
 
-	@Override
 	public UADEntityTypeComposite getUADEntityTypeComposite(
 		long userId, String key, int start, int end) {
 
@@ -164,7 +158,6 @@ public class UADRegistryImpl implements UADRegistry {
 			userId, key, getUADEntityDisplay(key), uadEntities);
 	}
 
-	@Override
 	public List<UADEntityTypeComposite> getUADEntityTypeComposites(
 		long userId, String uadEntitySetName) {
